requirements.txt
python-dotenv
gemini-api # Free AI model for script generation
requests   # API calls
moviepy    # Video editing
ffmpeg-python # Video automation
grpcio
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread 
genai
pandas
google-generativeai
google-cloud-texttospeech
huggingface_hub  
Pillow
selenium 
webdriver-manager